backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Google AI
google-generativeai==0.7.2
google-cloud-speech==2.26.0
google-cloud-texttospeech==2.16.3
google-cloud-firestore==2.16.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Twilio
twilio==9.0.4

# Utils
python-dotenv==1.0.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3
structlog==24.1.0
